==> RecSys2020Tutorial/03_2_Categorify.ipynb <==
# The MIT License (MIT)

# Copyright (c) 2020, NVIDIA CORPORATION.

# Permission is hereby granted, free of charge, to any person obtaining a copy of
# this software and associated documentation files (the "Software"), to deal in
# the Software without restriction, including without limitation the rights to
# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
# the Software, and to permit persons to whom the Software is furnished to do so,
# subject to the following conditions:

# The above copyright notice and this permission notice shall be included in all
# copies or substantial portions of the Software.

# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE

# Tutorial: Feature Engineering for Recommender Systems

# 3. Feature Engineering - Categorical

## 3.2. Categorify

import IPython

import pandas as pd
import cudf
import numpy as np
import cupy
import matplotlib.pyplot as plt

df_train = cudf.read_parquet('./data/train.parquet')
df_valid = cudf.read_parquet('./data/valid.parquet')
df_test = cudf.read_parquet('./data/test.parquet')

df_train.head()

cat = 'product_id'

## Theory

<b>*Categorifying*</b> is required for using categorical features in deep learning models with Embedding layers. An Embedding layer encodes the category into a hidden latent vector with a smaller dimension.<br><br>
Categorical features can be from datatype String or Integer. The Embedding layer requires that categorical features are continoues, positive Integers from 0 to |C| (number of unique category values).

There are 164453 unique product values but the ProductIDs range from 1000894 to 100144608.

df_train[cat].unique()

Using factorize creates continous Integers from a categorical fateature.

codes, uniques = df_train[cat].factorize()

codes

codes.unique()

Another important reason to Categorify categorical features is to reduce the size of the dataset. Often categorical features are of the datatype String and sometimes, they are hashed to protect the user / dataset privacy.

import hashlib
from sys import getsizeof

For example, we can hash the Integer 0 to a md5 hash

hashlib.md5(b'0').hexdigest()

We can hash the full product_id column

hashSeries = df_train[cat].to_pandas().apply(lambda x: hashlib.md5(bytes(str(x), encoding='utf-8')).hexdigest())

hashSeries

getsizeof(hashSeries)

codes, uniques = hashSeries.factorize()

getsizeof(pd.DataFrame(codes)[0])

We require only 9% of the original DataSeries memory.

91691016/1020060933

Finally, we can prevent overfitting for low frequency categories. Categories with low frequency can be grouped together to an new category called 'other'. In the previous exercise we learned that it is powerful to combine categorical features together to create a new feature. However, combining categories increases the cardinality of the new feature and the number of obersations per category will decrease. Therefore, we can apply a treshhold to group all categories with lower frequency count to the the new category.<br><br>
In addition, categories, which occure in the validation dataset and do not occur in the trainint dataset, should be mapped to the 'other' category as well.<br><br>
We use in our example the categoryIds 0 or 1 for a placeholder for the low frequency and unkown category. Then our function is independent of the cardinality of the categorical feature and we do not keep records of the cardinality to know the low frequency/unkown category.

In our dataset, we see that multiple product_ids occure only once in the training dataset. Our model would overfit to these low frequent categories.

df_train[cat].value_counts()

freq = df_train[cat].value_counts()

freq = freq.reset_index()
freq.columns = [cat, 'count']
freq = freq.reset_index()
freq.columns = [cat + '_Categorify', cat, 'count']
freq_filtered = freq[freq['count']>5]
freq_filtered[cat + '_Categorify'] = freq_filtered[cat + '_Categorify']+1
freq_filtered = freq_filtered.drop('count', axis=1)
df_train = df_train.merge(freq_filtered, how='left', on=cat)
df_train[cat + '_Categorify'] = df_train[cat + '_Categorify'].fillna(0)

df_train['product_id_Categorify'].min(), df_train['product_id_Categorify'].max(), df_train['product_id_Categorify'].drop_duplicates().shape

We need to apply the categorify to our validation and test sets.

df_valid = df_valid.merge(freq_filtered, how='left', on=cat)
df_valid[cat + '_Categorify'] = df_valid[cat + '_Categorify'].fillna(0)

df_test = df_test.merge(freq_filtered, how='left', on=cat)
df_test[cat + '_Categorify'] = df_test[cat + '_Categorify'].fillna(0)

#### Summary

<li> Categorify is important to enable deep learning models to use categorical features<br>
<li> Categorify can significantly reduce the dataset size by tranforming categorical features from String datatypes to Integer datatypes<br>
<li> Categorify can prevent overfitting by grouping categories with low frequency into one category together

## Practice

Now, it is your turn

**ToDo:**
<li> Categorify the category features brand
<li> Apply a frequency treshhold of minimum 20
<li> Map low frequency categories to the id=0
<li> Map unkown categories to the id=1 in the validation and test set

    
**Question:**
<li> How many data points have an unknown category in the test dataset?
<li> How many data points have a low frequency category in the test dataset?
<li> How many data points have a low frequency category in the training dataset?

### ToDo

############### Solution ###############

############### Solution End ###########

## Optimization

Let's compare the runtime between pandas and cuDF. The implementation depends only on the DataFrame object (calling function of the object) and does not require any pd / cuDF function. Therefore, we can use the same implementation and just use pandas.DataFrame and cuDF.DataFrame. 

def categorify(df_train, df_valid, df_test, cat, freq_treshhold=20, unkown_id=1, lowfrequency_id=0):
    freq = df_train[cat].value_counts()
    freq = freq.reset_index()
    freq.columns = [cat, 'count']
    freq = freq.reset_index()
    freq.columns = [cat + '_Categorify', cat, 'count']
    freq[cat + '_Categorify'] = freq[cat + '_Categorify']+2
    freq.loc[freq['count']<freq_treshhold, cat + '_Categorify'] = lowfrequency_id
    
    freq = freq.drop('count', axis=1)
    df_train = df_train.merge(freq, how='left', on=cat)
    df_train[cat + '_Categorify'] = df_train[cat + '_Categorify'].fillna(unkown_id)
    
    df_valid = df_valid.merge(freq, how='left', on=cat)
    df_valid[cat + '_Categorify'] = df_valid[cat + '_Categorify'].fillna(unkown_id)
    
    df_test = df_test.merge(freq, how='left', on=cat)
    df_test[cat + '_Categorify'] = df_test[cat + '_Categorify'].fillna(unkown_id)

df_train_pd = df_train.to_pandas()
df_valid_pd = df_valid.to_pandas()
df_test_pd = df_test.to_pandas()

%%time

categorify(df_train_pd, df_valid_pd, df_test_pd, 'user_id')

%%time

categorify(df_train, df_valid, df_test, 'user_id')

In our experiments, running the same implementation is 63x times faster with cuDF instead of pandas.

We shutdown the kernel.

app = IPython.Application.instance()
app.kernel.do_shutdown(False)
